Reorganise application structure => move/rename code
