+++ client/src/app/services/user.service.ts
@@ -6,16 +6,19 @@ import { AuthSuccessResponse, LoginCredentialsDTO, SignupCredentialsDTO } from '
 import { BaseHttpClient } from './base-http-client.service';
 import { LocalStorageService } from './local-storage.service';
 import { SocketService } from './socket.service';
+import { HotToastService } from '@ngneat/hot-toast';
+import { UserSearchResult } from 'src/shared/index.model';
 
 @Injectable({
     providedIn: 'root',
 })
-export class AuthService {
+export class UserService {
     constructor(
         private http: BaseHttpClient,
         private localStorageService: LocalStorageService,
         private store: Store,
         private socket: SocketService,
+        private toastService: HotToastService,
     ) {}
 
     // TODO: this should be implemented with effects
@@ -30,7 +33,7 @@ export class AuthService {
             })
             .catch(err => err);
     }
-    
+
     // TODO: and also this should be implemented with effects
     async signup(credentials: SignupCredentialsDTO): Promise<AuthSuccessResponse | HttpServerErrorResponse> {
         return this.http
@@ -51,5 +54,13 @@ export class AuthService {
     }
     loadUser() {
         return this.localStorageService.getUser();
+    }
+
+    searchUsers(query: string) {
+        if (!query) {
+            this.toastService.warning('You cannot search for an empty username.');
+            return;
+        }
+        return this.http.getAsync<UserSearchResult[]>(`/user/search?q=${query}`);
     }
 }
