+++ client/src/app/store/user/user.effects.ts
@@ -3,7 +3,7 @@ import { of } from 'rxjs';
 import { map, tap } from 'rxjs/operators';
 
 import { Injectable } from '@angular/core';
-import { AuthService } from 'src/app/services/auth.service';
+import { UserService } from 'src/app/services/user.service';
 import { userActions } from './user.actions';
 import { appActions } from '../app.actions';
 import { Router } from '@angular/router';
@@ -13,7 +13,7 @@ import { HotToastService } from '@ngneat/hot-toast';
 export class UserEffects {
     constructor(
         private actions$: Actions,
-        private authService: AuthService,
+        private userService: UserService,
         private router: Router,
         private toastService: HotToastService,
     ) {}
@@ -37,7 +37,7 @@ export class UserEffects {
                 tap(() => {
                     this.toastService.info('You logged out.');
                     this.router.navigate(['/auth/login']);
-                    this.authService.logout();
+                    this.userService.logout();
                 }),
             ),
         { dispatch: false },
@@ -48,7 +48,7 @@ export class UserEffects {
             ofType('@ngrx/effects/init'),
             map(() => {
                 console.log('loading user...');
-                const user = this.authService.loadUser();
+                const user = this.userService.loadUser();
 
                 if (user) return userActions.loadUserSuccess(user);
                 return appActions.nothing();
