+++ server/src/users/friendships/friendships.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { FriendshipsService } from './friendships.service';
+
+describe('FriendshipsService', () => {
+    let service: FriendshipsService;
+
+    beforeEach(async () => {
+        const module: TestingModule = await Test.createTestingModule({
+            providers: [FriendshipsService],
+        }).compile();
+
+        service = module.get<FriendshipsService>(FriendshipsService);
+    });
+
+    it('should be defined', () => {
+        expect(service).toBeDefined();
+    });
+});
